# 8292698: Improve performance of DataInputStream