Clean backport to improve `DataInputStream` performance in JDK 17u.

Benchmark from the changeset:

```
Benchmark                     Mode  Cnt   Score   Error  Units

#### x86_64

# Before
DataInputStreamTest.readChar  avgt   20  19.939 ? 0.120  us/op
DataInputStreamTest.readInt   avgt   20  20.538 ? 0.052  us/op

# After
DataInputStreamTest.readChar  avgt   20  10.497 ? 0.079  us/op
DataInputStreamTest.readInt   avgt   20   5.399 ? 0.073  us/op

### AArch64

# Before
DataInputStreamTest.readChar  avgt   20  21.600 ? 0.410  us/op
DataInputStreamTest.readInt   avgt   20  21.425 ? 0.545  us/op

# After
DataInputStreamTest.readChar  avgt   20  12.186 ? 0.366  us/op
DataInputStreamTest.readInt   avgt   20   6.207 ? 0.174  us/op
```

Additional testing:
 - [x] Benchmark added by RFE
 - [x] Linux x86_64 fastdebug, `tier1 tier2 tier3`
 - [x] Linux AArch64 fastdebug, `tier1 tier2 tier3`

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue

### Issue
 * [JDK-8292698](https://bugs.openjdk.org/browse/JDK-8292698): Improve performance of DataInputStream


### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.org/jdk17u-dev.git pull/1390/head:pull/1390` \
`$ git checkout pull/1390`

Update a local copy of the PR: \
`$ git checkout pull/1390` \
`$ git pull https://git.openjdk.org/jdk17u-dev.git pull/1390/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 1390`

View PR using the GUI difftool: \
`$ git pr show -t 1390`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.org/jdk17u-dev/pull/1390.diff">https://git.openjdk.org/jdk17u-dev/pull/1390.diff</a>

</details>


### Webrev
[Link to Webrev Comment](https://git.openjdk.org/jdk17u-dev/pull/1390#issuecomment-1559657559)